allow for custom hover.method()
