+++ lib/layer/_layer.mjs
@@ -2,7 +2,7 @@ import decorate from './decorate.mjs'
 
 import formats from './format/_format.mjs';
 
-import hover from './hover.mjs'
+import featureHover from './featureHover.mjs'
 
 import featureFilter from './featureFilter.mjs'
 
@@ -17,7 +17,7 @@ import graduated from './themes/graduated.mjs'
 export default {
   decorate,
   formats,
-  hover,
+  featureHover,
   featureFilter,
   Style,
   themes: {
