+++ lib/layer/decorate.mjs
@@ -97,8 +97,9 @@ export default async layer => {
   }
 
   if (layer.style?.hover) {
-
-    if (!layer.style.hover.show) mapp.layer.hover(layer);
+    
+    // Assign default featureHover method if non is provided.
+    layer.style.hover.method ??= mapp.layer.featureHover;
   }
 
   // Layer style has multiple themes.
