+++ lib/mapview/interactions/highlight.mjs
@@ -197,8 +197,8 @@ export default function (params) {
     if (mapview.interaction.current.layer.infoj) mapview.Map.getTargetElement().style.cursor = 'pointer'
 
     // Execute hover method assigned to the current feature layer.
-    typeof mapview.interaction.current.layer.style?.hover?.show === 'function'
-      && mapview.interaction.current.layer.style.hover.show(feature.F)
+    typeof mapview.interaction.current.layer.style?.hover?.method === 'function'
+      && mapview.interaction.current.layer.style.hover.method(feature.F, mapview.interaction.current.layer)
 
     // Style the feature itself if possible.
     if (mapview.interaction.current.layer.format !== 'mvt'
