+++ lib/ui/layers/panels/style.mjs
@@ -131,7 +131,8 @@ export default layer => {
           // Set hover from dropdown option.
           layer.style.hover = layer.style.hovers[entry.option]
   
-          if (!layer.style.hover.show) mapp.layer.hover(layer);
+          // Assign default featureHover method if non is provided.
+          layer.style.hover.method ??= mapp.layer.featureHover;
   
           layer.style.hover.display = display
         }
